requirements.txt
argh==0.28.1 
asgiref==3.7.2  
Django==4.2.3  
djangorestframework==3.14.0 
mysqlclient==2.2.0  
pip==23.1.2 
PyDispatcher==2.0.7
pytz==2023.3
PyYAML==6.0
setuptools==58.1.0
sqlparse==0.4.4
tool==0.8.0
typing_extensions==4.7.1
tzdata==2023.3